Moved color stuff from ConsoleDriver to Color.cs
